requirements.txt
attrs==21.4.0
importlib-metadata==4.8.3
iniconfig==1.1.1
packaging==21.3
pkg_resources==0.0.0
pluggy==1.0.0
PodSixNet==0.11.0
psutil==5.9.1
py==1.11.0
pyparsing==3.0.9
pytest==7.0.1
pytest-xprocess==0.18.1
tomli==1.2.3
typing_extensions==4.1.1
zipp==3.6.0
